Ejercicio modificado

+++ src/Sistema.java
@@ -1,9 +1,9 @@
 public class Sistema {
     public String nombre;
     public String materia;
-    public float nota;
+    public double nota;
 
-    public Sistema(String nombre, String materia, float nota) {
+    public Sistema(String nombre, String materia, double nota) {
         this.nombre = nombre;
         this.materia = materia;
         this.nota = nota;
@@ -25,11 +25,11 @@ public class Sistema {
         this.materia = materia;
     }
 
-    public float getNota() {
+    public double getNota() {
         return nota;
     }
 
-    public void setNota(float nota) {
+    public void setNota(double nota) {
         this.nota = nota;
     }
     public void mostrarDatos(){
